Split test class into two.
