+++ MetaphysicsIndustries.Solus.Test/FunctionsT/SizeFunctionT/CheckArgumentsTest.cs
@@ -28,7 +28,7 @@ using NUnit.Framework;
 namespace MetaphysicsIndustries.Solus.Test.FunctionsT.SizeFunctionT
 {
     [TestFixture]
-    public class SizeFunctionTest
+    public class CheckArgumentsTest
     {
         [Test]
         public void CheckArgumentsZeroArgumentsThrows()
@@ -114,58 +114,6 @@ namespace MetaphysicsIndustries.Solus.Test.FunctionsT.SizeFunctionT
             // expect
             Assert.DoesNotThrow(
                 () => SizeFunction.Value.CheckArguments(args));
-        }
-
-        private static SolusEnvironment empty = new SolusEnvironment();
-
-        [Test]
-        public void VectorYieldsLength()
-        {
-
-            // given
-            var args = new IMathObject[]
-            {
-                new float[] {1, 2}.ToVector()
-            };
-            // when
-            var result = SizeFunction.Value.Call(empty, args);
-            // then
-            Assert.IsTrue(result.IsScalar(null));
-            Assert.AreEqual(2, result.ToFloat());
-        }
-
-        [Test]
-        public void MatrixYieldsRowAndColumnCounts()
-        {
-            // given
-            var args = new IMathObject[]
-            {
-                new Matrix(new float[,]
-                {
-                    {1, 2, 3},
-                    {4, 5, 6}
-                })
-            };
-            // when
-            var result = SizeFunction.Value.Call(empty, args);
-            // then
-            Assert.IsTrue(result.IsVector(null));
-            var v = result.ToVector();
-            Assert.AreEqual(2, v.Length);
-            Assert.AreEqual(2, v[0].ToFloat());
-            Assert.AreEqual(3, v[1].ToFloat());
-        }
-
-        [Test]
-        public void StringYieldsLength()
-        {
-            // given
-            var args = new IMathObject[] {"abc".ToStringValue()};
-            // when
-            var result = SizeFunction.Value.Call(empty, args);
-            // then
-            Assert.IsTrue(result.IsScalar(null));
-            Assert.AreEqual(3, result.ToFloat());
         }
     }
 }
