+++ MetaphysicsIndustries.Solus.Test/MetaphysicsIndustries.Solus.Test.csproj
@@ -201,8 +201,9 @@
     <Compile Include="FunctionsT\SineFunctionT\GetResultTest.cs" />
     <Compile Include="FunctionsT\SineFunctionT\SineFunctionTest.cs" />
     <Compile Include="FunctionsT\SingleArgumentFunctionT\SingleArgumentFunctionTest.cs" />
+    <Compile Include="FunctionsT\SizeFunctionT\CheckArgumentsTest.cs" />
     <Compile Include="FunctionsT\SizeFunctionT\GetResultTest.cs" />
-    <Compile Include="FunctionsT\SizeFunctionT\SizeFunctionTest.cs" />
+    <Compile Include="FunctionsT\SizeFunctionT\CallTest.cs" />
     <Compile Include="FunctionsT\TangentFunctionT\GetResultTest.cs" />
     <Compile Include="FunctionsT\TangentFunctionT\TangentFunctionTest.cs" />
     <Compile Include="FunctionsT\UnaryOperationT\UnaryOperationTest.cs" />
